Changed the implementation of ThreadTransactionManager
Changed the implementation of `ThreadTransactionManager` to be a
`thread.local` that wraps a `TransactionManager` rather than a
`thread.local` that inherits from `TransactionManager`.  It now
exposes a `manager` attribute that allows access to the wrapped
transaction manager to allow cross thread calls.  ZODB will use this
to call `unregisterSynch` when a connection is closed from a main
thread.