Fix documentation of args parameter
From the documentation I got the wrong impression that the the hooks adder methods would take arguments as in `*args`.
But this is not the case as internally `tuple(args)` is called. This works for strings of length 1 as expected but it is an edge case which should not be stressed in the documentation.